Added tests for cases when the company lists
should not be displayed.
